pyproject.toml
[tool.poetry]
name = "big-ai-team"
version = "0.1.0"
description = "2026年AI智能体架构项目"
authors = ["Your Name <your.email@example.com>"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
redis = "^5.0.0"
alembic = "^1.13.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.24.0"
httpx = "^0.25.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
agno = "^2.5.14"
aiohttp = "^3.13.5"
gitpython = "^3.1.46"
tenacity = "^8.5.0"
psutil = "^7.2.2"
prometheus-client = "^0.24.1"
loguru = "^0.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"
bandit = "^1.7.6"
safety = "^2.3.5"
ruff = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unreachable = true
warn_return_any = true
warn_unused_configs = true
